requirements.txt
# requirements.txt
pandas
scikit-learn
joblib
PyYAML
dvclive
dvc
scikit-image
matplotlib 
seaborn
ipykernel
